requirements_interior.txt
# Interior Design AI - Pix2Pix with LoRA
# FastAPI Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Deep Learning
torch==2.1.0
diffusers==0.24.0
transformers==4.35.2
accelerate==0.25.0
safetensors==0.4.1
peft==0.7.1
huggingface-hub==0.19.4

# Image Processing
Pillow==10.1.0

# NumPy (compatible version)
numpy<2.0
